Bring in pest as a submodule
To make use of no_std, we bought in pest as a submodule under the
thirdparty directory. We also had to move the Cargo.toml to the
crates folder because pest itself has a workspace. Rust doesn't
support nested workspaces.

Several changes are made to build scripts to accomodate the change
to folder structure. `no_std` is back in the lib.rs file.